继续移植

+++ main/src/web_server.h
@@ -18,7 +18,7 @@ extern char g_ap_pass[64];
 extern struct fingerprint_device zw111; // Fingerprint device instance
 extern char g_touch_password[TOUCH_PASSWORD_LEN + 1];             // Current password
 extern uint64_t g_card_id_value[MAX_CARDS];
-extern int g_card_count;
+extern uint8_t g_card_count;
 
 httpd_handle_t web_server_start(void);
